Allow controllers to decide whether to assign attributes
This solves an issue where automatically assigning attributes would cause associations to be updated immediately if `_ids` attributes were assigned. This would be a problem if we later failed to save the rest of the attribute changes due to a validation error, resulting in some of the changes having been applied, while others had not.

Instead, we allow the controller to define which actions should automatically assign attributes by passing an `assign_attributes` option to `load_and_authorize_resource`. Passing false disables the automatic assignment for all actions, while passing one or more action name symbols will enable it for those actions.